requirements.txt
pytest
pytest-xdist
selenium==3.141.0
allure-pytest
mimesis
xlrd==1.2.0
os



